+++ lib/data/models/badge_model.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class BadgeModel {
+  final String? badgeKey;
+  final Color? badgeColor;
+  final Color? labelColor;
+  final TextStyle? labelStyle;
+
+  BadgeModel({
+    this.badgeKey,
+    this.badgeColor,
+    this.labelColor,
+    this.labelStyle,
+  });
+}