+++ lib/views/widgets/app_badge.dart
@@ -3,13 +3,17 @@ import 'package:flutter/material.dart';
 /// A simple badge widget to display counts.
 class AppBadge extends StatelessWidget {
   final int? count;
-  final Color color;
+  final Color? color;
+  final Color? labelColor;
+  final TextStyle? labelStyle;
   final double size;
 
   const AppBadge({
     super.key,
     required this.count,
-    this.color = Colors.red,
+    this.color,
+    this.labelColor,
+    this.labelStyle,
     this.size = 16.0,
   });
 
@@ -20,7 +24,7 @@ class AppBadge extends StatelessWidget {
     return Container(
       padding: const EdgeInsets.all(4),
       decoration: BoxDecoration(
-        color: color,
+        color: color ?? Colors.red,
         borderRadius: BorderRadius.circular(size),
       ),
       constraints: BoxConstraints(
@@ -30,11 +34,12 @@ class AppBadge extends StatelessWidget {
       child: Center(
         child: Text(
           count!.toString(),
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 10,
-            fontWeight: FontWeight.bold,
-          ),
+          style: labelStyle ??
+              TextStyle(
+                color: labelColor ?? Colors.white,
+                fontSize: 10,
+                fontWeight: FontWeight.bold,
+              ),
           textAlign: TextAlign.center,
         ),
       ),